+++ lib/dlms-cosem/include/dlms/cosem/simple_objects.hpp
@@ -1002,6 +1002,58 @@ private:
   CosemAccessRights rights_;
 };
 
+class CosemAutoConnectObject : public ICosemObject
+{
+public:
+  static const std::uint8_t MaxSupportedVersion = 0u;
+
+  CosemAutoConnectObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& mode,
+    const CosemByteBuffer& repetitions,
+    const CosemByteBuffer& repetitionDelay,
+    const CosemByteBuffer& callingWindow,
+    const CosemByteBuffer& destinationList,
+    AttributeAccessMode mutableAccess);
+  CosemAutoConnectObject(
+    const CosemLogicalName& logicalName,
+    const CosemByteBuffer& mode,
+    const CosemByteBuffer& repetitions,
+    const CosemByteBuffer& repetitionDelay,
+    const CosemByteBuffer& callingWindow,
+    const CosemByteBuffer& destinationList,
+    AttributeAccessMode mutableAccess,
+    std::uint8_t version);
+
+  CosemObjectDescriptor Descriptor() const;
+  CosemAccessRights AccessRights() const;
+  CosemStatus ReadAttribute(
+    std::uint8_t attributeId,
+    CosemByteBuffer& output) const;
+  CosemStatus WriteAttribute(
+    std::uint8_t attributeId,
+    const CosemByteBuffer& input);
+  CosemStatus InvokeMethod(
+    std::uint8_t methodId,
+    const CosemByteBuffer& input,
+    CosemByteBuffer& output);
+
+  const CosemByteBuffer& Mode() const;
+  const CosemByteBuffer& Repetitions() const;
+  const CosemByteBuffer& RepetitionDelay() const;
+  const CosemByteBuffer& CallingWindow() const;
+  const CosemByteBuffer& DestinationList() const;
+
+private:
+  CosemObjectDescriptor descriptor_;
+  CosemByteBuffer mode_;
+  CosemByteBuffer repetitions_;
+  CosemByteBuffer repetitionDelay_;
+  CosemByteBuffer callingWindow_;
+  CosemByteBuffer destinationList_;
+  CosemAccessRights rights_;
+};
+
 enum class CosemClockBase
 {
   NotDefined = 0,